+++ Sources/StreamChat/Models/Attachments/AttachmentTypes.swift
@@ -223,7 +223,7 @@ public enum AttachmentFileType: String, Codable, Equatable, CaseIterable {
     /// Video
     case mov, avi, wmv, webm
     /// Image
-    case jpeg, png, gif, bmp, webp
+    case jpeg, png, gif, bmp, webp, heic, heif, svg
     /// Unknown
     case unknown
 
@@ -260,7 +260,12 @@ public enum AttachmentFileType: String, Codable, Equatable, CaseIterable {
         "image/png": .png,
         "image/gif": .gif,
         "image/bmp": .bmp,
-        "image/webp": .webp
+        "image/webp": .webp,
+        "image/heic": .heic,
+        "image/heic-sequence": .heic,
+        "image/heif": .heif,
+        "image/heif-sequence": .heif,
+        "image/svg+xml": .svg
     ]
 
     /// Init an attachment file type by mime type.
@@ -306,6 +311,15 @@ public enum AttachmentFileType: String, Codable, Equatable, CaseIterable {
     public var isAudio: Bool {
         switch self {
         case .mp3, .wav, .ogg, .m4a, .aac:
+            return true
+        default:
+            return false
+        }
+    }
+
+    public var isImage: Bool {
+        switch self {
+        case .jpeg, .png, .webp, .gif, .bmp:
             return true
         default:
             return false